PY-129 Extract api utilities used in new code from util.py to internal/api_utils.py (#369)
